+++ packages/liveblocks-react-ui/src/primitives/AiMessage/types.ts
@@ -1,6 +1,8 @@
 import type {
+  AiAssistantMessage,
   AiChatMessage,
   AiReasoningPart,
+  AiRetrievalPart,
   AiTextPart,
   AiToolInvocationPart,
 } from "@liveblocks/core";
@@ -31,8 +33,17 @@ export type AiMessageContentToolInvocationPartProps = {
   /** @internal */
   isStreaming: boolean;
   /** @internal */
-  message: AiChatMessage;
+  message: AiAssistantMessage;
   part: AiToolInvocationPart;
+};
+
+/** @internal */
+export type AiMessageContentRetrievalPartProps = {
+  /** @internal */
+  index: number;
+  /** @internal */
+  isStreaming: boolean;
+  part: AiRetrievalPart;
 };
 
 export interface AiMessageContentComponents {
@@ -45,6 +56,11 @@ export interface AiMessageContentComponents {
    * The component used to display reasoning parts.
    */
   ReasoningPart: ComponentType<AiMessageContentReasoningPartProps>;
+
+  /**
+   * The component used to display knowledge retrieval parts.
+   */
+  RetrievalPart: ComponentType<AiMessageContentRetrievalPartProps>;
 
   /**
    * NOTE that ToolInvocationPart is slightly different.